add grpc server

+++ ecomm-grpc/storer/types.go
@@ -10,7 +10,7 @@ type Product struct {
 	Description  string     `db:"description"`
 	Rating       int64      `db:"rating"`
 	NumReviews   int64      `db:"num_reviews"`
-	Price        float64    `db:"price"`
+	Price        float32    `db:"price"`
 	CountInStock int64      `db:"count_in_stock"`
 	CreatedAt    time.Time  `db:"created_at"`
 	UpdatedAt    *time.Time `db:"updated_at"`
@@ -19,9 +19,9 @@ type Product struct {
 type Order struct {
 	ID            int64      `db:"id"`
 	PaymentMethod string     `db:"payment_method"`
-	TaxPrice      float64    `db:"tax_price"`
-	ShippingPrice float64    `db:"shipping_price"`
-	TotalPrice    float64    `db:"total_price"`
+	TaxPrice      float32    `db:"tax_price"`
+	ShippingPrice float32    `db:"shipping_price"`
+	TotalPrice    float32    `db:"total_price"`
 	UserID        int64      `db:"user_id"`
 	CreatedAt     time.Time  `db:"created_at"`
 	UpdatedAt     *time.Time `db:"updated_at"`
@@ -33,7 +33,7 @@ type OrderItem struct {
 	Name      string  `db:"name"`
 	Quantity  int64   `db:"quantity"`
 	Image     string  `db:"image"`
-	Price     float64 `db:"price"`
+	Price     float32 `db:"price"`
 	ProductID int64   `db:"product_id"`
 	OrderID   int64   `db:"order_id"`
 }
